Fix deletion of STI temp dir on exit
Code cleanup/refactoring.

Change call from os.Remove to os.RemoveAll to remove the temp dir and
its contents.

Only chown /tmp/artifacts.

Remove the chgrp portion of the chown calls as it shouldn't be necessary
and it keeps the user running sti from being able to delete the temp
dir.

Change /tmp/artifacts from 755 to 775 so it's group writable, meaning it
should be able to be deleted by the host user.

Fix issue where the build would fail when trying to determine the
default scripts URL if the image didn't already exist locally.

Fixes #164