Allow geard to run STI built images
Modify geard so it bind mounts a directory for the container init and
cmd scripts, instead of bind mounting the individual files. This is
temporary and can be changed back to files once we have a build of
Docker with
https://github.com/dotcloud/docker/commit/d535d9810012b18ca83f86aa416ff4a08bb49804
in it.

Change STI to put the container init script in /.sti.init so it doesn't
conflict with geard's /.container.init directory (this is also temporary
and will be changed to file bind mounts for the same reason as stated
above)